add files and display data
