+++ src/Components/Search.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+
+
+export default class Search extends Component {
+  constructor() {
+    super()
+    this.state = {
+    }
+  }
+  render() {
+    return(
+      <div>
+      <input placeholder="Search" />
+    </div>
+    )
+  }
+}
