+++ src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import { render } from 'react-dom';
-import './index.css';
-
-import App from './App';
+import './css/index.css';
+import App from './Components/App';
 
 render( <App />, document.getElementById('root') );
